rag-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
langchain==0.3.14
langchain-anthropic==0.3.4
langchain-weaviate==0.1.0
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.4
weaviate-client==4.10.4
sentence-transformers==3.3.1
musicbrainzngs==0.7.1
lyricsgenius==3.0.1
httpx==0.28.1
tenacity==9.0.0
pybreaker==1.2.0
